backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pypdf>=4.0.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.0
